add locale_helper.cpp
