+++ sa/CMakeLists.txt
@@ -114,6 +114,7 @@ set(BASE_SOURCES
   ${BASE_DIR}/cubrid_getopt_long.c
   ${BASE_DIR}/binaryheap.c
   ${BASE_DIR}/tsc_timer.c
+  ${BASE_DIR}/locale_helper.cpp
   ${BASE_DIR}/lock_free.c
   ${BASE_DIR}/lockfree_bitmap.cpp
   ${BASE_DIR}/lockfree_hashmap.cpp
